Server based group chat application using python
This project was created as a part of capstone project for first year. To run the application successfully, first run the serverfin.py file and then the clientfin.py file. 
The server and client have to be on the same device i.e. localhost and must be connected through the same wi-fi network. TCP model was used for the project.